% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insidesexp-package.R
\docType{package}
\name{insidesexp-package}
\alias{insidesexp}
\alias{insidesexp-package}
\title{insidesexp: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdsumner/insidesexp}
  \item Report bugs at \url{https://github.com/mdsumner/insidesexp/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
