% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insidesexp-package.R
\docType{package}
\name{insidesexp-package}
\alias{insidesexp}
\alias{insidesexp-package}
\title{insidesexp: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
